Fix venv update if not created by buildenv -- resolves #86

+++ buildenv.sh
@@ -10,7 +10,7 @@ else
 fi
 ${_BUILDENV_PYTHON} --version 2>/dev/null 1>&2
 if test $? -ne 0; then
-    echo ">> ERROR: ${_BUILDENV_PYTHON} is not installed"
+    echo "[ERROR] ${_BUILDENV_PYTHON} is not installed"
     exit 1
 fi
 
@@ -24,11 +24,13 @@ if test ${_BUILDENV_RC} -eq 100; then
     ${SHELL} --rcfile .buildenv/shell.sh
     _BUILDENV_RC=$?
 elif test ${_BUILDENV_RC} -gt 100; then
-    # Execute command if required
+    # Execute command if required (and command script found)
     _BUILDENV_CMD=.buildenv/command.${_BUILDENV_RC}.sh
-    ${SHELL} -c ${_BUILDENV_CMD}
-    _BUILDENV_RC=$?
-    rm ${_BUILDENV_CMD}
+    if test -f ${_BUILDENV_CMD}; then
+        ${SHELL} -c ${_BUILDENV_CMD}
+        _BUILDENV_RC=$?
+        rm ${_BUILDENV_CMD}
+    fi
 fi
 
 # Forward RC
